get macos icon show, but there's problems
It's not persistent, it's only while I have the app open, if I pin
it in the dock it won't have it, and also when I quit it, it reverts
to the default icon for a split second before it finishes quitting.

I don't know what to do, maybe we just have to involve Swift and a
proper macOS application.